Move OrloClient to separate file
